requirements.txt
google-genai
pydantic
jinja2
python-dotenv
boto3
pytest
aws-sam-cli
